更改 QQ_t 和 GID_t 类型实现

+++ include/defs/message_chain.hpp
@@ -52,7 +52,7 @@ namespace Cyan
 		{
 			json j;
 			j["type"] = "At";
-			j["target"] = qq;
+			j["target"] = int64_t(qq);
 			j["display"] = "@@";
 			messages_.push_back(j);
 			return *this;
